address

+++ Entities/User.cs
@@ -10,7 +10,7 @@ namespace Loza.Entities
 
         public string LastName { get; set; }
 
-        public int Wallet { get; set; } = 0;
+        public decimal Wallet { get; set; } = 0;
         
         public string PhoneNumber { get; set; }
